protect campground edit and delete routes

+++ routes/campgrounds.js
@@ -63,6 +63,10 @@ router.get(
       req.flash('error', 'Cannot find that campground');
       return res.redirect('/campgrounds');
     }
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You are not authorized to do that!');
+      return res.redirect(`/campgrounds/${id}`);
+    }
     res.render('campgrounds/edit', { campground });
   })
 );
@@ -72,7 +76,12 @@ router.put(
   validateCampground,
   catchAsync(async (req, res) => {
     const id = req.params.id;
-    const campground = await Campground.findByIdAndUpdate(id, {
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You are not authorized to do that!');
+      return res.redirect(`/campgrounds/${id}`);
+    }
+    const camp = await Campground.findByIdAndUpdate(id, {
       ...req.body.campground,
     });
     req.flash('success', 'Successfully updated campground!');
@@ -83,7 +92,12 @@ router.delete(
   '/:id',
   isLoggedIn,
   catchAsync(async (req, res) => {
-    const { id } = req.params;
+    const id = req.params.id;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You are not authorized to do that!');
+      return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted campground!');
     res.redirect('/campgrounds');
